requirements.txt
absl-py==1.3.0
accelerate==0.14.0
adabelief-pytorch==0.2.1
addict==2.4.0
aim==3.15.2
aim-ui==3.15.2
aimrecords==0.0.7
aimrocks==0.2.1
aiofiles==22.1.0
aiohttp==3.8.3
aiosignal==1.2.0
albumentations==0.5.2
alembic==1.9.1
aliyun-log-python-sdk==0.8.2
altair==5.0.1
antlr4-python3-runtime==4.9.3
anyio==3.7.1
astor==0.8.1
asttokens==2.2.1
async-timeout==4.0.2
asynctest==0.13.0
attrs==22.1.0
axial-positional-embedding==0.2.1
backcall==0.2.0
backports.zoneinfo==0.2.1
base58==2.0.1
bcrypt==4.0.1
beautifulsoup4==4.11.1
black==22.3.0
blobfile==2.0.2
boto3==1.26.83
botocore==1.29.83
braceexpand==0.1.7
cachetools==5.2.0
Cerberus==1.3.4
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cloths-segmentation==0.0.2
cloudpickle==2.2.0
colorama==0.4.6
coloredlogs==15.0.1
common==0.1.2
contrastive-learner==0.1.1
cryptography==39.0.0
cupy==6.0.0
cycler==0.11.0
dateparser==1.1.7
debugpy==1.6.6
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.13
detectron2==0.6
dill==0.3.6
docker-pycreds==0.4.0
dominate==2.7.0
easydict==1.10
efficientnet-pytorch==0.7.1
einops==0.6.0
elastic-transport==8.4.0
elasticsearch==8.6.2
exceptiongroup==1.1.2
executing==1.2.0
fairscale==0.4.6
fastapi==0.67.0
fastjsonschema==2.15.3
fastrlock==0.8.1
ffmpy==0.3.1
filelock==3.8.0
fire==0.5.0
Flask==2.2.2
flatbuffers==1.12
fonttools==4.38.0
frozenlist==1.3.1
fsspec==2022.10.0
future==0.18.2
fvcore==0.1.5.post20220512
gast==0.2.2
gdown==4.6.0
gitdb==4.0.10
GitPython==3.1.29
google-auth==2.13.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
gradio==3.34.0
gradio_client==0.2.6
greenlet==2.0.1
grpcio==1.50.0
h11==0.14.0
h5py==3.7.0
httpcore==0.17.3
httpx==0.24.1
huggingface-hub==0.16.4
humanfriendly==10.0
hydra-core==1.3.0.dev0
icecream==2.1.3
idna==3.4
iglovikov-helper-functions==0.0.53
imagecorruptions==1.1.2
imageio==2.22.2
imgaug==0.4.0
importlib-metadata==5.0.0
importlib-resources==5.10.0
iopath==0.1.9
ipython==7.34.0
itchat==1.2.32
itsdangerous==2.1.2
jedi==0.19.0
Jinja2==3.0.0
jmespath==1.0.1
joblib==1.2.0
jpeg4py==0.1.4
jsonschema==4.17.3
jstyleson==0.0.2
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
kornia==0.5.4
lightning-utilities==0.5.0
linear-attention-transformer==0.19.1
linformer==0.2.1
linkify-it-py==2.0.2
lltm-cpp==0.0.0
lltm-cuda==0.0.0
lmdb==1.4.0
local-attention==1.7.0
lpips==0.1.4
lxml==4.9.3
lz4==4.0.2
Mako==1.2.4
Markdown==3.4.1
markdown-it-py==2.2.0
MarkupSafe==2.1.1
matplotlib==3.5.3
matplotlib-inline==0.1.6
mdit-py-plugins==0.3.3
mdurl==0.1.2
megfile==2.0.1
mpmath==1.2.1
mprop==0.16.0
multidict==6.0.2
munch==2.5.0
mypy-extensions==0.4.3
networkx==2.6.3
ninja==1.10.2.4
numpy==1.21.6
oauthlib==3.2.2
omegaconf==2.3.0.dev1
onnx==1.11.0
onnx-simplifier==0.4.17
onnxoptimizer==0.3.8
onnxruntime==1.14.1
onnxruntime-training==1.14.1
onnxsim==0.4.17
opencv-python==4.6.0.66
opencv-python-headless==4.6.0.66
openvino==2022.3.0
openvino-dev==2022.3.0
openvino-telemetry==2022.3.0
opt-einsum==3.3.0
orjson==3.9.2
packaging==21.3
paddle-bfloat==0.1.7
paddlelite==2.12
pandas==1.3.5
paramiko==3.0.0
parso==0.8.3
pathspec==0.10.1
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
piq==0.7.0
pkgutil_resolve_name==1.3.10
platformdirs==2.5.2
portalocker==2.6.0
prefetch-generator==1.0.1
pretrainedmodels==0.7.4
prettytable==3.5.0
product-key-memory==0.1.10
promise==2.3
prompt-toolkit==3.0.39
protobuf==3.20.0
psutil==5.9.4
ptyprocess==0.7.0
py==1.11.0
py3nvml==0.2.7
pyarrow==11.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybind11==2.6.2
pycocotools==2.0.5
pycparser==2.21
pycryptodomex==3.18.0
pydantic==1.10.4
pyDeprecate==0.3.2
pydot==1.4.2
pydub==0.25.1
Pygments==2.14.0
PyNaCl==1.5.0
pyparsing==3.0.9
pypng==0.20220715.0
PyQRCode==1.2.1
pyrsistent==0.19.3
PySocks==1.7.1
python-dateutil==2.8.2
python-multipart==0.0.6
pytorch-fid==0.3.0
pytorch-lightning==1.8.6
pytorch-msssim==0.2.1
pytorch-toolbelt==0.5.2
pytz==2022.5
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.3.0
PyYAML==5.4.1
qudida==0.0.4
regex==2022.10.31
requests==2.28.1
requests-oauthlib==1.3.1
RestrictedPython==6.0
retry==0.9.2
rich==13.3.1
rsa==4.9
s3transfer==0.6.0
scikit-image==0.16.1
scikit-learn==1.0.2
scipy==1.7.3
seaborn==0.12.1
segmentation-models-pytorch==0.3.1
semantic-version==2.10.0
sentry-sdk==1.11.1
setproctitle==1.3.2
sewar==0.4.5
Shapely==1.8.5.post1
shortuuid==1.0.11
shutup==0.2.0
six==1.16.0
smart-open==6.3.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.46
starlette==0.14.2
stylegan2-pytorch==1.8.9
sympy==1.10.1
tabulate==0.9.0
tensorboard==1.14.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.5.1
tensorflow-estimator==1.14.0
tensorflow-gpu==1.14.0
termcolor==2.1.0
texttable==1.6.7
threadpoolctl==3.1.0
tifffile==2021.11.2
timm==0.4.12
tomli==2.0.1
toolz==0.12.0
torch-ema==0.3
torchmetrics==0.10.1
torchscan==0.1.2
tqdm==4.64.1
traitlets==5.9.0
typed-ast==1.5.4
typing_extensions==4.4.0
tzdata==2022.7
tzlocal==4.2
uc-micro-py==1.0.2
urllib3==1.26.12
uvicorn==0.20.0
vector-quantize-pytorch==0.1.0
waitress==2.1.2
wandb==0.13.6
wcwidth==0.2.5
webdataset==0.2.26
websockets==11.0.3
Werkzeug==2.2.2
wrapt==1.14.1
wxpy==0.3.9.8
x2paddle==1.4.0
xmltodict==0.13.0
yacs==0.1.8
yarl==1.8.1
zipp==3.10.0
